changed Integer to Long to reflect changes in the db (int -> bigint)

+++ src/main/java/com/misiac/workoutjournal/repository/WorkoutRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(exported = false)
-public interface WorkoutRepository extends JpaRepository<Workout, Integer> {
+public interface WorkoutRepository extends JpaRepository<Workout, Long> {
 }
